requirements.txt
flask==2.3.3
gunicorn==21.2.0
flask-cors==4.0.0
numpy==1.24.3
pandas==1.5.3
openpyxl==3.1.2
werkzeug==2.3.7
requests==2.31.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
flask-login==0.6.2
email-validator==2.0.0
flask-limiter==3.12
cryptography==41.0.7
selenium==4.15.2
flask-wtf==1.2.1
redis==5.0.1
